requirements.txt
pydantic
starlette
uvicorn
fastapi
python-multipart
python-dotenv
typing-extensions
anyio
loguru
langgraph
zhipuai
mcp
PyPDF2
pdfminer.six
openpyxl>=3.1.2
jinja2>=3.1.2
